applications: initial release
