+++ beard/applications/learning_model.py
@@ -1,0 +1,221 @@
+# -*- coding: utf-8 -*-
+#
+# This file is part of Beard.
+# Copyright (C) 2015 CERN.
+#
+# Beard is a free software; you can redistribute it and/or modify it
+# under the terms of the Revised BSD License; see LICENSE file for
+# more details.
+
+"""Tools for learning the distance model."""
+
+import cPickle
+import json
+import numpy as np
+
+from sklearn.ensemble import GradientBoostingClassifier
+from sklearn.feature_extraction.text import CountVectorizer
+from sklearn.feature_extraction.text import TfidfVectorizer
+from sklearn.pipeline import FeatureUnion
+from sklearn.pipeline import Pipeline
+
+from beard.applications.utils import get_author_full_name
+from beard.applications.utils import get_author_other_names
+from beard.applications.utils import get_author_initials
+from beard.applications.utils import get_author_affiliation
+from beard.applications.utils import get_title
+from beard.applications.utils import get_journal
+from beard.applications.utils import get_abstract
+from beard.applications.utils import get_coauthors
+from beard.applications.utils import get_keywords
+from beard.applications.utils import get_collaborations
+from beard.applications.utils import get_references
+from beard.applications.utils import get_year
+from beard.applications.utils import group_by_signature
+from beard.applications.utils import load_signatures
+
+from beard.similarity import AbsoluteDifference
+from beard.similarity import CosineSimilarity
+from beard.similarity import PairTransformer
+from beard.utils import FuncTransformer
+from beard.utils import Shaper
+
+
+def _build_distance_estimator(X, y):
+    """Build a vector reprensation of a pair of signatures."""
+    transformer = FeatureUnion([
+        ("author_full_name_similarity", Pipeline([
+            ("pairs", PairTransformer(element_transformer=Pipeline([
+                ("full_name", FuncTransformer(func=get_author_full_name)),
+                ("shaper", Shaper(newshape=(-1,))),
+                ("tf-idf", TfidfVectorizer(analyzer="char_wb",
+                                           ngram_range=(2, 4),
+                                           dtype=np.float32,
+                                           decode_error="replace")),
+            ]), groupby=group_by_signature)),
+            ("combiner", CosineSimilarity())
+        ])),
+        ("author_other_names_similarity", Pipeline([
+            ("pairs", PairTransformer(element_transformer=Pipeline([
+                ("other_names", FuncTransformer(func=get_author_other_names)),
+                ("shaper", Shaper(newshape=(-1,))),
+                ("tf-idf", TfidfVectorizer(analyzer="char_wb",
+                                           ngram_range=(2, 4),
+                                           dtype=np.float32,
+                                           decode_error="replace")),
+            ]), groupby=group_by_signature)),
+            ("combiner", CosineSimilarity())
+        ])),
+        ("author_initials_similarity", Pipeline([
+            ("pairs", PairTransformer(element_transformer=Pipeline([
+                ("initials", FuncTransformer(func=get_author_initials)),
+                ("shaper", Shaper(newshape=(-1,))),
+                ("count", CountVectorizer(analyzer="char_wb",
+                                          ngram_range=(1, 1),
+                                          binary=True,
+                                          decode_error="replace")),
+            ]), groupby=group_by_signature)),
+            ("combiner", CosineSimilarity())
+        ])),
+        ("affiliation_similarity", Pipeline([
+            ("pairs", PairTransformer(element_transformer=Pipeline([
+                ("affiliation", FuncTransformer(func=get_author_affiliation)),
+                ("shaper", Shaper(newshape=(-1,))),
+                ("tf-idf", TfidfVectorizer(analyzer="char_wb",
+                                           ngram_range=(2, 4),
+                                           dtype=np.float32,
+                                           decode_error="replace")),
+            ]), groupby=group_by_signature)),
+            ("combiner", CosineSimilarity())
+        ])),
+        ("coauthors_similarity", Pipeline([
+            ("pairs", PairTransformer(element_transformer=Pipeline([
+                ("coauthors", FuncTransformer(func=get_coauthors)),
+                ("shaper", Shaper(newshape=(-1,))),
+                ("tf-idf", TfidfVectorizer(dtype=np.float32,
+                                           decode_error="replace")),
+            ]), groupby=group_by_signature)),
+            ("combiner", CosineSimilarity())
+        ])),
+        ("title_similarity", Pipeline([
+            ("pairs", PairTransformer(element_transformer=Pipeline([
+                ("title", FuncTransformer(func=get_title)),
+                ("shaper", Shaper(newshape=(-1,))),
+                ("tf-idf", TfidfVectorizer(analyzer="char_wb",
+                                           ngram_range=(2, 4),
+                                           dtype=np.float32,
+                                           decode_error="replace")),
+            ]), groupby=group_by_signature)),
+            ("combiner", CosineSimilarity())
+        ])),
+        ("journal_similarity", Pipeline([
+            ("pairs", PairTransformer(element_transformer=Pipeline([
+                ("journal", FuncTransformer(func=get_journal)),
+                ("shaper", Shaper(newshape=(-1,))),
+                ("tf-idf", TfidfVectorizer(analyzer="char_wb",
+                                           ngram_range=(2, 4),
+                                           dtype=np.float32,
+                                           decode_error="replace")),
+            ]), groupby=group_by_signature)),
+            ("combiner", CosineSimilarity())
+        ])),
+        ("abstract_similarity", Pipeline([
+            ("pairs", PairTransformer(element_transformer=Pipeline([
+                ("abstract", FuncTransformer(func=get_abstract)),
+                ("shaper", Shaper(newshape=(-1,))),
+                ("tf-idf", TfidfVectorizer(dtype=np.float32,
+                                           decode_error="replace")),
+            ]), groupby=group_by_signature)),
+            ("combiner", CosineSimilarity())
+        ])),
+        ("keywords_similarity", Pipeline([
+            ("pairs", PairTransformer(element_transformer=Pipeline([
+                ("keywords", FuncTransformer(func=get_keywords)),
+                ("shaper", Shaper(newshape=(-1,))),
+                ("tf-idf", TfidfVectorizer(dtype=np.float32,
+                                           decode_error="replace")),
+            ]), groupby=group_by_signature)),
+            ("combiner", CosineSimilarity())
+        ])),
+        ("collaborations_similarity", Pipeline([
+            ("pairs", PairTransformer(element_transformer=Pipeline([
+                ("collaborations", FuncTransformer(func=get_collaborations)),
+                ("shaper", Shaper(newshape=(-1,))),
+                ("tf-idf", TfidfVectorizer(dtype=np.float32,
+                                           decode_error="replace")),
+            ]), groupby=group_by_signature)),
+            ("combiner", CosineSimilarity())
+        ])),
+        ("references_similarity", Pipeline([
+            ("pairs", PairTransformer(element_transformer=Pipeline([
+                ("references", FuncTransformer(func=get_references)),
+                ("shaper", Shaper(newshape=(-1,))),
+                ("tf-idf", TfidfVectorizer(dtype=np.float32,
+                                           decode_error="replace")),
+            ]), groupby=group_by_signature)),
+            ("combiner", CosineSimilarity())
+        ])),
+        ("year_diff", Pipeline([
+            ("pairs", FuncTransformer(func=get_year, dtype=np.int)),
+            ("combiner", AbsoluteDifference())
+        ]))])
+
+    # Train a classifier on these vectors
+    classifier = GradientBoostingClassifier(n_estimators=500,
+                                            max_depth=9,
+                                            max_features=10,
+                                            learning_rate=0.125,
+                                            verbose=3)
+
+    # Return the whole pipeline
+    estimator = Pipeline([("transformer", transformer),
+                          ("classifier", classifier)]).fit(X, y)
+
+    return estimator
+
+
+def learn_model(distance_pairs, input_signatures, input_records,
+                distance_model):
+    """Learn the distance model using the files from the disc.
+
+    Parameters
+    ----------
+    :param distance_pairs: string
+        Path to the file with distance pairs. The content should be a json
+        array of tuples (`signature_id1`, `signature_id2`, `target`),
+        where `target = 0` if both signatures belong to the same author,
+        and `target = 1` otherwise.
+
+        [(0, 1, 0), (2, 3, 0), (4, 5, 1), ...]
+
+    :param input_signatures: string
+        Path to the file with signatures. The content should be a json array
+        of dictionaries holding metadata about signatures.
+
+        [{"signature_id": 0,
+          "author_name": "Doe, John",
+          "publication_id": 10, ...}, { ... }, ...]
+
+    :param input_records: string
+        Path to the file with records. The content should be a json array of
+        dictionaries holding metadata about records
+
+        [{"publication_id": 0,
+          "title": "Author disambiguation using Beard", ... }, { ... }, ...]
+    """
+    pairs = json.load(open(distance_pairs, "r"))
+    signatures, records = load_signatures(input_signatures, input_records)
+
+    X = np.empty((len(pairs), 2), dtype=np.object)
+    y = np.empty(len(pairs), dtype=np.int)
+
+    for k, (i, j, target) in enumerate(pairs):
+        X[k, 0] = signatures[i]
+        X[k, 1] = signatures[j]
+        y[k] = target
+
+    # Learn a distance estimator on paired signatures
+    distance_estimator = _build_distance_estimator(X, y)
+    cPickle.dump(distance_estimator,
+                 open(distance_model, "w"),
+                 protocol=cPickle.HIGHEST_PROTOCOL)
