+++ beard/clustering/__init__.py
@@ -11,10 +11,12 @@
 
 from .blocking import BlockClustering
 from .blocking_funcs import block_double_metaphone
+from .blocking_funcs import block_last_name_first_initial
 from .blocking_funcs import block_single
 from .wrappers import ScipyHierarchicalClustering
 
 __all__ = ("BlockClustering",
            "block_double_metaphone",
+           "block_last_name_first_initial",
            "block_single",
            "ScipyHierarchicalClustering")
